+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## 0.9.0
+
+#### 💥 Breaking
+
+- Removed `install_global`, use `cargo install` instead.
+- Removed `uninstall_global`, use `cargo uninstall` instead.
+
+#### 🚀 Updates
+
+- Updated to support proto v0.31 release.
+- Updated dependencies.
+
 ## 0.8.1
 
 #### 🐞 Fixes